fix issue where some properties were not being matched for deserialization
Added tests for the rest of the files included in the caffe googlenet model. Upon adding them, it was noticed that in some cases, the .proto generators might just use the default name as it comes from the proto, instead of PascalCaseing it. This was primarily due to properties with their type being a nested enum with the same name as the property.